+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rustic_print"
-version = "0.1.1"
+version = "0.2.0"
 edition = "2021"
 description = "A versatile Rust library for enhancing console output. It offers a range of features to create a more engaging and informative command-line interface."
 license = "MIT"
@@ -13,4 +13,4 @@ categories = ["command-line-interface", "text-processing", "development-tools", 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-crossterm = "0.27"
+crossterm = "0.28.1"
